Creation des services - Definir variables environnement - Creation des DTO - ...

+++ src/main/java/ca/uqtr/inf1013_h22_g5/repository/PharmacienRepository.java
@@ -4,6 +4,5 @@ import ca.uqtr.inf1013_h22_g5.model.Pharmacien;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface PharmacienRepository extends JpaRepository<Pharmacien, Long> {
-
+public interface PharmacienRepository extends JpaRepository<Pharmacien, Integer> {
 }
